reduced counter implementation to 32bit

+++ src/impl/vrEmuTms9918Priv.h
@@ -98,9 +98,9 @@ struct vrEmuTMS9918_s
   uint8_t palWriteStage;
   uint8_t palWriteStage0Value;
 
-  uint64_t startTime;
-  uint64_t stopTime;
-  uint64_t currentTime;
+  uint32_t startTime;
+  uint32_t stopTime;
+  uint32_t currentTime;
 
   bool scanlineHasSprites;
 };
